refactor sync code
